Time: 1 ms (88.85%), Space: 42.7 MB (61.14%) - LeetHub

+++ 0050-powx-n/0050-powx-n.java
@@ -1,7 +1,7 @@
 class Solution {
     public double myPow(double x, int n) {
         double ans=1.0;
-        long nn=n;
+        double nn=n;
         if(nn<0) nn=-1*nn;
         while(nn>0)
         {
